+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update Jest snapshots
 
 4.15.0 - (October 3, 2019)
 ------------------