+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `isInvalid` prop and stylings
 
 4.15.0 - (October 3, 2019)
 ------------------